cleanup comments, use database table instead of wrappers
