+++ lib/runtime/runtime.go
@@ -29,6 +29,18 @@ import (
 var memory, memErr = wasm.NewMemory(17, 0)
 var logger = log.New("pkg", "runtime")
 
+// NodeStorageTypePersistent flog to identify offchain storage as persistent (db)
+const NodeStorageTypePersistent int32 = 1
+
+// NodeStorageTypeLocal flog to identify offchain storage as local (memory)
+const NodeStorageTypeLocal int32 = 2
+
+// NodeStorage struct for storage of runtime offchain worker data
+type NodeStorage struct {
+	LocalStorage      BasicStorage
+	PersistentStorage BasicStorage
+}
+
 // Ctx struct
 type Ctx struct {
 	storage     Storage
